fix: final fixes
