+++ language/translations/en.json
@@ -1351,5 +1351,7 @@
   "Hero cards are a summary of your activity in the Bitcoin space. You can share them with your friends!": "Hero cards are a summary of your activity in the Bitcoin space. You can share them with your friends!",
   "Creator Ranking": "Creator Ranking",
   "Contributor Ranking": "Contributor Ranking",
-  "Ambassador Ranking": "Ambassador Ranking"
+  "Ambassador Ranking": "Ambassador Ranking",
+  "Current month": "Current month",
+  "Current week": "Current week"
 }
