+++ src/modules/profile/pages/profilePage/views/account/views/UserBio.tsx
@@ -1,5 +1,4 @@
 import { SkeletonText, VStack } from '@chakra-ui/react'
-import { t } from 'i18next'
 
 import { useUserProfileAtom } from '@/modules/profile/state'
 import { Body } from '@/shared/components/typography'
@@ -15,9 +14,6 @@ export const UserBio = () => {
 
   return (
     <VStack w="full" alignItems={'start'}>
-      <Body size="xl" medium>
-        {t('Bio')}
-      </Body>
       <Body>{userProfile.bio}</Body>
     </VStack>
   )
@@ -26,9 +22,6 @@ export const UserBio = () => {
 export const UserBioSkeleton = () => {
   return (
     <VStack w="full" alignItems={'start'}>
-      <Body size="xl" medium>
-        {t('Bio')}
-      </Body>
       <Body>
         <SkeletonText noOfLines={3} spacing="4" />
       </Body>
